evaluation/requirements.txt
azure-ai-evaluation
pandas
python-dotenv
openai
